Create README.md
It’s a simple application for showcasing movies. This application uses The Movie Database (TMDb) API to get the information
from and show them into two sections where you can view them into the ”popular movies” and ”top-rated movies”. Using the
API, the application fetches the movie list and shows the poster of the movies into these categories so that the app can mimic
cineplex feeling. When the user touches on a movie, it will open new activity and will show the description of that movie and
show the rating of that movie. A poster of that movie is there also. A search bar is there to let the user search for a movie.
Because I attempted this for the first time to develop some Android applications, I thought of designing something I will be using
daily. While designing, I thought of those functionalities, which will make my movie selection easy, and I tried to will get the
best out of my experience in movie searching. Developing a good functional movie showcasing app requires great expertise and
practice. Because of my limited knowledge and practice, I succeeded in implementing only a few features into the app, and there
is a lot of room for improvement,